Implement aes encryption helper in crypto crate
