+++ Cargo.toml
@@ -32,3 +32,4 @@ axum-extra = { version = "0.8.0" }
 axum-server = { version = "0.5", features = ["tls-rustls"] }
 rustls = { version = "0.23", features = ["ring"] }
 reqwest = { version = "0.12", features = ["json", "rustls-tls"] }
+crypto = { path = "./packages/crypto" }
